Big refactor
- Use templating.
- Use separate stylesheet.
- Rename backends->backends.logging (there will be other backends).
- Create server package.
