+++ mnist_numpy/server/__main__.py
@@ -1,0 +1,26 @@
+import click
+import uvicorn
+from loguru import logger
+
+from mnist_numpy.server.app import DB_PATH
+
+
+@click.command()
+@click.option("--host", default="127.0.0.1", help="Host to bind to")
+@click.option("--port", default=8000, help="Port to bind to")
+@click.option("--reload", is_flag=True, help="Enable auto-reload")
+def main(host: str, port: int, reload: bool):
+    logger.info(
+        f"Starting MNIST Training Monitor on http://{host}:{port} using database: {DB_PATH}"
+    )
+    uvicorn.run(
+        "mnist_numpy.server.app:app",
+        host=host,
+        port=port,
+        reload=reload,
+        log_level="info",
+    )
+
+
+if __name__ == "__main__":
+    main()
