+++ mnist_numpy/server/static/css/dashboard.css
@@ -1,0 +1,2 @@
+/* Dashboard-specific styles could go here if needed */
+/* For now, all styles are in the main styles.css */
