+++ mnist_numpy/templates/base.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>{% block title %}MNIST Training Monitor{% endblock %}</title>
+    <link rel="stylesheet" href="{{ url_for('static', path='/css/base.css') }}">
+    <link rel="stylesheet" href="{{ url_for('static', path='/css/components.css') }}">
+    {% block extra_css %}{% endblock %}
+</head>
+<body>
+    <div class="container">
+        <div class="header">
+            <h1>{% block header_title %}MNIST Training Monitor{% endblock %}</h1>
+            <p>{% block header_subtitle %}Real-time visualisation of neural network training progress{% endblock %}</p>
+        </div>
+        
+        {% block nav %}
+        <div class="nav-bar">
+            <div class="nav-item"><strong>Current Run:</strong> <span id="current-run">Loading...</span></div>
+            <div class="nav-links">
+                <a href="/latest" class="nav-link">Latest</a>
+                <a href="/manage" class="nav-link">Manage</a>
+                {% block extra_nav %}{% endblock %}
+            </div>
+        </div>
+        {% endblock %}
+        
+        {% block content %}{% endblock %}
+    </div>
+    
+    {% block scripts %}{% endblock %}
+</body>
+</html>
