+++ mnist_numpy/templates/dashboard.html
@@ -1,0 +1,53 @@
+{% extends "base.html" %}
+
+{% block title %}Dashboard - MNIST Training Monitor{% endblock %}
+
+{% set plotly = true %}
+
+{% block extra_nav %}
+<select id="run-selector" class="run-dropdown" onchange="navigateToRun()">
+    <option value="">Select Run...</option>
+</select>
+{% endblock %}
+
+{% block extra_css %}
+<link rel="stylesheet" href="{{ url_for('static', path='/css/dashboard.css') }}">
+{% endblock %}
+
+{% block content %}
+<button class="refresh-btn" onclick="refreshData()">Refresh Data</button>
+<div id="status" class="status"></div>
+<div class="stats-grid" id="stats-grid"></div>
+
+<div class="plot-container">
+    <h3>Training Progress - Loss Over Epochs</h3>
+    <div id="loss-plot" style="width:100%;height:500px;"></div>
+</div>
+
+<div class="plot-container">
+    <h3>Learning Rate Schedule</h3>
+    <div id="lr-plot" style="width:100%;height:400px;"></div>
+</div>
+
+<div class="plot-container">
+    <h3>Training Timeline</h3>
+    <div id="timeline-plot" style="width:100%;height:400px;"></div>
+</div>
+{% endblock %}
+
+{% block scripts %}
+<script>
+// Your existing JavaScript here - much cleaner to read!
+async function fetchData() {
+    try {
+        const response = await fetch('/api/data');
+        return response.ok ? await response.json() : null;
+    } catch (error) {
+        console.error('Error fetching data:', error);
+        return null;
+    }
+}
+
+// ... rest of your JS
+</script>
+{% endblock %}
