+++ pyproject.toml
@@ -73,4 +73,4 @@ resume = "mnist_numpy.cli:resume"
 sample_data = "mnist_numpy.cli:sample_data"
 trace_statistics = "mnist_numpy.scripts.trace_statistics:main"
 train = "mnist_numpy.cli:train"
-serve_plot = "mnist_numpy.scripts.serve_plot:main"
+serve = "mnist_numpy.server.__main__:main"
